Fix based on the code reviewer's suggestions: - Remove all unnecessary comments. - Modify the Acknowledgments of the README file; thanks to the designer. - Separate the testing into small testing files corresponding to components.
